+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
 import lotto.domain.Lotto;
